Fix compiling error and some clippy warnings. (#61)
* Fix compile and some clippies
* Replace tab to 4 space
* Align naming

+++ examples/top_trap_with_collisions.rs
@@ -129,7 +129,7 @@ fn main() {
         dispatcher.dispatch(&mut world);
         world.maintain();
 
-        if (_i > 0) && (_i % 50 as i32 == 0) {
+        if (_i > 0) && (_i % 50_i32 == 0) {
             let tracker = world.read_resource::<CollisionsTracker>();
             let _result = write_collisions_tracker(
                 &mut filename,
